Prepare web30 for full block requests
This patch fixes TransactionResponse parsing to accept contract creation
events and adds a test case for that.

It also adds a request size tuning parameter that makes it viable to
request full blocks that may be filled with contract creations or other
large operations.

Finally we change the default params of actix-web to use compression by
default