net_version ergonomics
net_version is always a number across all json_rpc implementations
yet it's returned as a string by our implementation. This patch
adds automating parsing in the net_version endpoint.

In addition to that send_transaction abstracts gas and many other
things about sending a transaction but not the network ID if it's
going to do that it should either make providing the network ID
mandatory or retrieve it itself. This patch simply has it automatically
aquire the network ID.

There is a security impact from this second chain. The device many
now be tricked into signing a transaction for any chain the full node
desires. I do not think this is too serious of a problem for most users.
For those specifically building cross chain applications they can read
the doc comment and act accordingly.