Improve event API
The event parsing functions where up until this point pretty immature.
This patch should provide a more flexible API with minimal impact
on ease of use and an increase in saftey and clarity of operations.

Changes:
* Use a flexible rather than a fixed number of topics, previously it
was actually impossible to find events with less than N indexed topics
where N was the number of provided filter arguments. This hidden gotcha
is now resolved.
* Provide a flexible number of addresses to watched, while this wasn't
causing strange issues it is more limiting than the actual log api and
vec![] is easy enough to use for me to judge it worth exposing
* No more NewFilter::default() which is just asking for someone to forget
to fill a field and have strange function behavior
* Provide a meaningful behavior difference between wait_for_event() and
wait_for_event_alt() where the latter is one call but will always wait
the provided time whereas the former will make more requests but may exit
early