Modify events interface to take a vector
This completes the refactor started in 0.8.0 to remove the fixed
number of event arguments and replace it with a more compact vec
interface.

I think there may be practical limits to the number of arguments that
can be passed in this vector, but if that's the case I'll add in a
error when a vector that is too long is used.