Better specify min Clarity version
This has been causing strange build issues across the board requiring a
manual Clarity crate update. The issue is that new elements of the
Clarity API where added, but the new minimum version was not specified
here. So it will only compile if you manually update it. This release
corrects that issue.

It also bumps the dev dependencies to their latest versions.