contract_call -> simulate_transaction
This patch bumps Web30 for 0.15.0 and renames the contract_call function
to simulate_transaction. This better represents what the tx actually
does, and is paired with a much expanded doc comment detailing how the
function can be used and it's limiations.

The other major chang ehere is in how the TransactionRequest is
formatted, previously it was formatted very differently from the actual
transaction. Skipping the many from trait implementations handled in
Clarity for direct wrapping. I suspec this as contributing to encoding
flaws.