Use /proc/meminfo to size memory allocation
This patch removes the ability for the user to hint at the desired
response size on ETH calls. This value was not all that useful and
created problems around incorrect hints.

What this patch does is set a default buffer size to an enormous 10GB,
this will not OOM users in the common case because memory is only
allocated as actually needed, a few tens of megabytes at most.

The primary concern to be resolved in this patch was dealing with
OpenWRT routers and other low memory systems where it's very feasible to
have a 50MB response OOM the device on deserialization. So we import a
very simple 'proc/meminfo' parser from althea_rs that lets us set the
maximum parsing size to a fraction of the system available memory and
ensure that we can both parse the largest responses possible and we
incur no crash or OOM risk.

Due to the way this is implemented it extends to almost all Linux
systems and provides protection for potential 500MB or 1GB response
sizes on say small cloud systems as well.