Add wait_timeout to wrapping and amm functions
Prevously the functions for wrapping ETH and making AMM calls did not
have the standard wait_timeout parameter used for functions that make
transactions.

This was not revealed in testing becuase the use of HardHat on automine
caused transactions to be instantly included, bypassing the the
real-world issue of race conditions caused by the block queue process.

When a tx is submitted it takes some time to confirm, during that time
if you send another tx it's nonce will be in conflict with the first tx,
since eth_get_transaction_count() does not include pending transactions.

What this does in effect is create a pileup of race conditions in the
code anywhere transactions are sent but not waited for.