Check approved in swap_uniswap
Previously this function would always send an approval tx without
checking if it was really needed or not. Furthermore when sending that
appoval tx it would not wait for the response, meaning that the
following transaction would share the same txid and barring some race
conditions the trade would execute and fail without the approval or the
approval would execute and the trade never run.

This patch resolves these issues by manually specifying the nonce,
allowing both transactions to enter the mempool and run sequentially, we
also check if approval is actually required.