Added a syncing node check for functions in web30
Several function here depend on the fact the our eth node is synced and upto date. For instance, the function get_balances
checks balances and when this function is called without our node being synced, we can get behaviors such as low balance
notification when this isnt the case. This commit fixes this issue by added a check to see if our node is synced first