Bump for 0.18.3
The main goal here is to deal with the way post eip1559 blockchains deal
with fees, in situations where fees are not limiting (xDai) eth_gasPrice
is inflated and not required, in sitautions where fees are limited
EIP1559 takes over and the gas price is effectively the base gas price.
Either way we shouldn't use eth_gasPrice unless on a pre-london chain.