Add generic gas estimator
This is based off of an existing gas price estimator written by Pranay
in althea_rs I'm in need of a very similar structure for the Gravity
Bridge relayer so I figured a generic gas price storage and estimation
system would be useful.

The general challenge here is trying to balance everyones risk and
requirements while remaining simple. This is a first draft and will be
iterated as it's used.

Co-authored-by: Pranay Tulugu <pranay@althea.systems>