Use eth_gas_price by default
In the past we have tried to optimize gas by using the block base fee
but this has turned out to be fraught with issues because the base fee
often goes up and becomes insufficient in the next block. This changes
the default gas price to eth_gas_price rather than the base fee.