Add erc20 balance fetching for contracts with 0 ETH
When fetching erc20 balances, the `target_address` was used as the gas
payer, but that meant querying the Gravity.sol balance resulted in an
InsufficientGas error.

get_erc20_balance_as_address and get_erc20_balance_at_height_as_address
have been added to address this issue.

This is a non-breaking API change, functionality of the existing API has
not changed.